backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
web3==6.11.3
ipfshttpclient==0.8.0a2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
alembic==1.12.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-slugify==8.0.1
cryptography==41.0.8
hashlib3==2.0.1